requirements.txt
antspyx==0.5.4
certifi==2024.12.14
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
etelemetry==0.3.1
filelock==3.17.0
fonttools==4.55.6
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
kiwisolver==1.4.8
looseversion==1.3.0
lxml==5.3.0
matplotlib==3.10.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.8.6
numpy==2.0.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
prov==2.0.1
pydot==3.0.4
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rdflib==6.3.2
requests==2.32.3
scipy==1.15.1
simplejson==3.19.3
six==1.17.0
statsmodels==0.14.4
traits==6.3.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webcolors==24.11.1
